fix(db): make the PostgreSQL half read and write what the repository decodes
The PostgreSQL dialect had never been executed — no server was ever installed,
so every second statement was written and unverified. This runs it, for the
first time, against PostgreSQL 17.11, and fixes what that found.

Eight defects, every one a violation of ADR 0004's rule that each parameter and
each selected column is String or i64 so rows decode identically on both
engines:

* set_password_hash bound one parameter list for two statements whose
  placeholders are in different orders — the insert leads with account_id, the
  update cannot, because SET comes before WHERE. SQLite accepted the surplus
  parameter and shifted every value by one, so the WHERE compared an account id
  against a timestamp, matched no row, and the call returned Ok(()) having
  written nothing. A password change did nothing and reported success, and every
  test in the tree passed. PostgreSQL refused it outright, which is how it was
  found. This was a live bug on SQLite, not only on PostgreSQL.
* ON CONFLICT DO UPDATE with no conflict target, in privacy_settings,
  outbox_events and reading_history_entry. Legal in SQLite, refused by
  PostgreSQL; the privacy and progress statements carry partial unique indexes,
  so the target names the index and repeats its predicate.
* version INTEGER against i64 in Rust, and the same for twenty more integer
  columns. SQLite's INTEGER is 64-bit and PostgreSQL's is 32-bit, so those
  columns are BIGINT now — ADR 0004 amended.
* the 0/1 flags stay BOOLEAN and are read as bool::int::bigint, because
  PostgreSQL has no boolean-to-bigint cast.
* a ? missing its ::uuid cast in the rating upsert.
* reading_progress_unique was declared without device_id on PostgreSQL, so two
  devices reading one work would have collided onto a single row. That is a
  divergence between the two migration trees, which are checked for identical
  ids and for nothing else.
* library_items bound four values for five placeholders.

The regression test is
lorehaven-db::identity::tests::setting_a_password_twice_replaces_the_stored_credential,
which fails on the old binding.

+++ migrations/postgres/0001_identity.sql
@@ -17,7 +17,7 @@ CREATE TABLE accounts (
     age_state         TEXT    NOT NULL DEFAULT 'unknown',
     created_at        TEXT    NOT NULL,
     updated_at        TEXT    NOT NULL,
-    version           INTEGER NOT NULL DEFAULT 1,
+    version           BIGINT NOT NULL DEFAULT 1,
     deleted_at        TEXT
 );
 
@@ -77,7 +77,7 @@ CREATE TABLE pseuds (
     discoverability TEXT    NOT NULL DEFAULT 'listed',
     created_at      TEXT    NOT NULL,
     updated_at      TEXT    NOT NULL,
-    version         INTEGER NOT NULL DEFAULT 1,
+    version         BIGINT NOT NULL DEFAULT 1,
     deleted_at      TEXT
 );
